Moved hull calculations to Hull constrcutor, moved Edge to it's own file

+++ src/ca/gkelly/engine/graphics/Camera.java
@@ -225,7 +225,7 @@ public class Camera {
 	 * @param y The y coordinate, in screen space
 	 * @return The coordinates, in world space
 	 */
-	public int[] worldSpace(int x, int y) {
+	public int[] worldSpace(double x, double y) {
 		int newX = (int) ((x - this.x) / zoom);
 		int newY = (int) ((y - this.y) / zoom);
